# docs/swagger.yaml
basePath: /
definitions:
...
  address.resolveResponse:
    properties:
      address_info:
        $ref: '#/definitions/location.AddressInfo'
      display_address:
        type: string
      query:
        type: string
      resolver:
        type: string
    type: object
...
  location.AddressInfo:
    properties:
      city:
        type: string
      country:
        type: string
      postal_code:
        type: string
      state_or_province:
        type: string
      street:
        type: string
      street_number:
        type: string
      unit:
        type: string
    type: object
...
info:
  contact: {}
  description: Address and place information service.
  title: Gopos
paths:
...
  /address/resolve:
    get:
      parameters:
      - description: Address to resolve.
        in: query
        name: query
        required: true
        type: string
      - description: Country of the address, improves accuracy if known.
        in: query
        name: country
        type: string
      - default: libpostal
        description: Resolver to use.
        enum:
        - libpostal
        - google
        in: query
        name: resolver
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/address.resolveResponse'
        "400":
          description: Missing or invalid query parameters.
      summary: Resolve an address query into structured address information.
...
swagger: "2.0"